+++ CHANGELOG.md
@@ -85,6 +85,7 @@
     changes to the agreements only affect new users
 - **Profile Changes**
   - Users can now delete their account
+  - Users can now directly upload profile pictures
 - **Gem Changes**
   - Updated to ruby 2.4.2
   - Updated to rails 5.1.4